Add SUES UC2-express-bus pinout as a struct
Maps the desired functionality to ESP32 pins that support it, and those pins have been made a specification for the SUES module-based system.
I made some variable names up or changed them. Now all of the GPIOs of the SUES system (see the document linked in the code comment) are in the struct.